requirements.txt
flask
ultralytics
opencv-python
SpeechRecognition
pydub
ffmpeg-python
torch
torchvision